src/worker/requirements.txt
# ── Storage adapters ──────────────────────────────────────────────────────────
minio==7.2.0
influxdb-client==1.38.0

# ── Messaging ─────────────────────────────────────────────────────────────────
paho-mqtt==1.6.1

# ── Data processing ───────────────────────────────────────────────────────────
pandas==2.1.3
numpy==1.26.2

# ── Configuration ─────────────────────────────────────────────────────────────
python-dotenv==1.0.0

# ── Monitoring ────────────────────────────────────────────────────────────────
prometheus-client==0.19.0

# ── HTTP ──────────────────────────────────────────────────────────────────────
requests==2.31.0

# ── Structured logging ────────────────────────────────────────────────────────
python-json-logger==2.0.7

# ── Data validation ───────────────────────────────────────────────────────────
pydantic==2.5.2

# ── Testing ───────────────────────────────────────────────────────────────────
pytest==7.4.3